+++ crates/tsz-emitter/src/transforms/class_es5_ast_to_ir.rs
@@ -8,7 +8,7 @@ use crate::context::transform::TransformDirective;
 use crate::transforms::async_es5_ir::AsyncES5Transformer;
 use rustc_hash::FxHashSet;
 use tsz_common::common::ModuleKind;
-use tsz_parser::parser::node::{FunctionData, Node};
+use tsz_parser::parser::node::FunctionData;
 use tsz_parser::parser::node_flags;
 use tsz_parser::syntax::transform_utils::{
     collect_class_computed_name_this_references, contains_new_target_reference,
@@ -1794,70 +1794,6 @@ impl<'a> AstToIr<'a> {
         }
         // Await expressions are handled by the async transform.
         IRNode::ASTRef(idx)
-    }
-
-    fn convert_type_assertion(&self, idx: NodeIndex) -> IRNode {
-        let node = self
-            .arena
-            .get(idx)
-            .expect("NodeIndex must be valid in arena");
-        // Both TYPE_ASSERTION and AS_EXPRESSION use TypeAssertionData
-        if let Some(assertion) = self.arena.get_type_assertion(node) {
-            let expression = self.convert_expression(assertion.expression);
-            if let Some(comment) = self.leading_block_comment_before_node(node).or_else(|| {
-                self.leading_block_comments_before_expression(node, assertion.expression)
-            }) {
-                IRNode::LeadingCommentExpr {
-                    comment: comment.into(),
-                    expression: Box::new(expression),
-                }
-            } else {
-                expression
-            }
-        } else {
-            IRNode::ASTRef(idx)
-        }
-    }
-
-    pub(super) fn leading_block_comments_before_expression(
-        &self,
-        node: &Node,
-        expression_idx: NodeIndex,
-    ) -> Option<String> {
-        let source_text = self.source_text?;
-        let expr_node = self.arena.get(expression_idx)?;
-        if node.pos >= expr_node.pos {
-            return None;
-        }
-        let start = node.pos as usize;
-        let end = (expr_node.pos as usize).min(source_text.len());
-        let mut comments = Vec::new();
-        for comment in tsz_common::comments::get_comment_ranges(source_text) {
-            let comment_start = comment.pos as usize;
-            let comment_end = comment.end as usize;
-            if comment_start >= end {
-                break;
-            }
-            if comment_start >= start && comment_end <= end && comment.is_multi_line {
-                comments.push(source_text[comment_start..comment_end].to_string());
-            }
-        }
-        (!comments.is_empty()).then(|| comments.join(" "))
-    }
-
-    pub(super) fn leading_block_comment_before_node(&self, node: &Node) -> Option<String> {
-        let source_text = self.source_text?;
-        let node_start = node.pos as usize;
-        let comment = tsz_common::comments::get_comment_ranges(source_text)
-            .into_iter()
-            .take_while(|comment| comment.end as usize <= node_start)
-            .filter(|comment| comment.is_multi_line)
-            .last()?;
-        let comment_end = comment.end as usize;
-        if comment_end > node_start || !source_text[comment_end..node_start].trim().is_empty() {
-            return None;
-        }
-        Some(source_text[comment.pos as usize..comment_end].to_string())
     }
 
     fn convert_non_null(&self, idx: NodeIndex) -> IRNode {
